+++ lib/src/cancel_upload_exception.dart
@@ -1,0 +1,10 @@
+class CancelUploadException implements Exception {
+  final String message;
+
+  const CancelUploadException([this.message]);
+
+  String toString() {
+    if (message == null) return 'CancelUploadException';
+    return 'CancelUploadException: $message';
+  }
+}