+++ lib/src/transport.dart
@@ -1,43 +1,76 @@
+import 'dart:async';
+import 'dart:io';
+
 import 'package:http/http.dart';
 import 'package:meta/meta.dart';
 import 'package:uploadcare_client/src/authorization/scheme.dart';
 
-class UcMultipartRequest extends MultipartRequest {
-  UcMultipartRequest({
-    @required String method,
-    @required Uri uri,
-    AuthScheme scheme,
-  })  : this._scheme = scheme,
-        super(method, uri) {
-    headers.addAll({'Content-Type': 'application/x-www-form-urlencoded'});
-  }
+mixin _AuthorizedRequestMixin on BaseRequest {
+  AuthScheme get scheme;
 
-  final AuthScheme _scheme;
-
-  @override
-  Future<StreamedResponse> send() {
-    if (_scheme != null) _scheme.authorizeRequest(this);
-
-    return super.send();
+  void authorize() {
+    if (scheme != null) scheme.authorizeRequest(this);
   }
 }
 
-class UcRequest extends Request {
+mixin _CancelableRequestMixin on BaseRequest, _AuthorizedRequestMixin {
+  final Client _client = Client();
+
+  void cancel() => _client.close();
+
+  @override
+  Future<StreamedResponse> send() async {
+    authorize();
+
+    try {
+      var response = await _client.send(this);
+      var stream = _onDone(response.stream, _client.close);
+      return StreamedResponse(
+        ByteStream(stream),
+        response.statusCode,
+        contentLength: response.contentLength,
+        request: response.request,
+        headers: response.headers,
+        isRedirect: response.isRedirect,
+        persistentConnection: response.persistentConnection,
+        reasonPhrase: response.reasonPhrase,
+      );
+    } catch (_) {
+      _client.close();
+      rethrow;
+    }
+  }
+
+  Stream<T> _onDone<T>(Stream<T> stream, void onDone()) =>
+      stream.transform(StreamTransformer.fromHandlers(handleDone: (sink) {
+        sink.close();
+        onDone();
+      }));
+}
+
+class UcMultipartRequest extends MultipartRequest
+    with _AuthorizedRequestMixin, _CancelableRequestMixin {
+  UcMultipartRequest({
+    @required String method,
+    @required Uri uri,
+    this.scheme,
+  }) : super(method, uri) {
+    headers.addAll(
+        {HttpHeaders.contentTypeHeader: 'application/x-www-form-urlencoded'});
+  }
+
+  final AuthScheme scheme;
+}
+
+class UcRequest extends Request
+    with _AuthorizedRequestMixin, _CancelableRequestMixin {
   UcRequest({
     @required String method,
     @required Uri uri,
-    AuthScheme scheme,
-  })  : _scheme = scheme,
-        super(method, uri) {
-    headers.addAll({'Content-Type': 'application/json'});
+    this.scheme,
+  }) : super(method, uri) {
+    headers.addAll({HttpHeaders.contentTypeHeader: 'application/json'});
   }
 
-  final AuthScheme _scheme;
-
-  @override
-  Future<StreamedResponse> send() {
-    if (_scheme != null) _scheme.authorizeRequest(this);
-
-    return super.send();
-  }
+  final AuthScheme scheme;
 }